requirements.txt
fastapi
uvicorn[standard]
duckdb
httpx
python-dotenv
google-generativeai
python-multipart
pydantic
pypdf
python-docx
